+++ server/src/main/java/org/allaymc/server/blockentity/type/AllayBlockEntityType.java
@@ -1,7 +1,7 @@
 package org.allaymc.server.blockentity.type;
 
+import com.google.common.base.Preconditions;
 import lombok.Getter;
-import lombok.SneakyThrows;
 import me.sunlan.fastreflection.FastConstructor;
 import org.allaymc.api.blockentity.BlockEntity;
 import org.allaymc.api.blockentity.component.BlockEntityComponent;
@@ -14,6 +14,7 @@ import org.allaymc.server.component.interfaces.ComponentProvider;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Supplier;
@@ -27,7 +28,6 @@ public class AllayBlockEntityType<T extends BlockEntity> implements BlockEntityT
     @Getter
     protected String name;
 
-    @SneakyThrows
     protected AllayBlockEntityType(Function<BlockEntityInitInfo, T> instanceCreator, String name) {
         this.instanceCreator = instanceCreator;
         this.name = name;
@@ -37,7 +37,6 @@ public class AllayBlockEntityType<T extends BlockEntity> implements BlockEntityT
         return new Builder(clazz);
     }
 
-    @SneakyThrows
     @Override
     public T createBlockEntity(BlockEntityInitInfo info) {
         info.setBlockEntityType(this);
@@ -45,7 +44,6 @@ public class AllayBlockEntityType<T extends BlockEntity> implements BlockEntityT
     }
 
     public static class Builder {
-
         protected Class<?> clazz;
         protected Map<Identifier, ComponentProvider<? extends BlockEntityComponent>> componentProviders = new HashMap<>();
         protected String name;
@@ -59,46 +57,45 @@ public class AllayBlockEntityType<T extends BlockEntity> implements BlockEntityT
             return this;
         }
 
-        public Builder setComponents(Map<Identifier, ComponentProvider<? extends BlockEntityComponent>> componentProviders) {
-            if (componentProviders == null)
-                throw new BlockEntityTypeBuildException("Component providers cannot be null");
-            this.componentProviders = new HashMap<>(componentProviders);
+        public Builder setComponents(Map<Identifier, ComponentProvider<? extends BlockEntityComponent>> providers) {
+            Preconditions.checkNotNull(providers, "providers");
+            this.componentProviders = new HashMap<>(providers);
             return this;
         }
 
-        public Builder addComponents(Map<Identifier, ComponentProvider<? extends BlockEntityComponent>> componentProviders) {
-            this.componentProviders.putAll(componentProviders);
+        public Builder addComponents(Map<Identifier, ComponentProvider<? extends BlockEntityComponent>> providers) {
+            this.componentProviders.putAll(providers);
             return this;
         }
 
         public Builder addComponent(Function<BlockEntityInitInfo, ? extends BlockEntityComponent> provider, Class<?> componentClass) {
-            var p = new ComponentProvider.SimpleComponentProvider<>(provider, componentClass);
+            var p = new ComponentProvider.Simple<>(provider, componentClass);
             this.componentProviders.put(p.findComponentIdentifier(), p);
             return this;
         }
 
         public Builder addComponent(Supplier<? extends BlockEntityComponent> supplier, Class<?> componentClass) {
-            return addComponent(info -> supplier.get(), componentClass);
+            return addComponent($ -> supplier.get(), componentClass);
         }
 
-        public Builder addComponent(ComponentProvider<? extends BlockEntityComponent> p) {
-            this.componentProviders.put(p.findComponentIdentifier(), p);
+        public Builder addComponent(ComponentProvider<? extends BlockEntityComponent> provider) {
+            this.componentProviders.put(provider.findComponentIdentifier(), provider);
             return this;
         }
 
         public <T extends BlockEntity> BlockEntityType<T> build() {
+            Preconditions.checkNotNull(name, "name");
             if (!componentProviders.containsKey(BlockEntityBaseComponentImpl.IDENTIFIER)) {
                 addComponent(BlockEntityBaseComponentImpl::new, BlockEntityBaseComponentImpl.class);
             }
-            if (name == null) throw new BlockEntityTypeBuildException("name cannot be null!");
 
             Function<BlockEntityInitInfo, T> instanceCreator;
             try {
-                var fastMemberConstructor = FastConstructor.create(clazz.getConstructors()[0]);
+                var constructor = FastConstructor.create(clazz.getConstructor(BlockEntityInitInfo.class, List.class));
                 var componentProviderList = new ArrayList<>(componentProviders.values());
                 instanceCreator = info -> {
                     try {
-                        return (T) fastMemberConstructor.invoke(info, componentProviderList);
+                        return (T) constructor.invoke(info, componentProviderList);
                     } catch (Throwable t) {
                         throw new BlockEntityTypeBuildException("Failed to create block entity instance!", t);
                     }