+++ server/src/main/java/org/allaymc/server/item/component/food/ItemRottenFleshFoodComponentImpl.java
@@ -8,8 +8,8 @@ import java.util.concurrent.ThreadLocalRandom;
 /**
  * @author IWareQ
  */
-public class ItemRottenFleshBaseComponentImpl extends ItemFoodComponentImpl {
-    public ItemRottenFleshBaseComponentImpl() {
+public class ItemRottenFleshFoodComponentImpl extends ItemFoodComponentImpl {
+    public ItemRottenFleshFoodComponentImpl() {
         super(4, 0.8f);
     }