+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Fixed random bounds.
 - Fixed `BlockBreakEvent` cancelling.
 - Fixed spam click bug.
+- Fixed chicken food points.
 
 ## 0.3.0 (API 0.6.0) - 2025-3-28